app/fileupload_event_service_read_pubsub/requirements.txt
google-cloud-pubsub
google-cloud-storage
pandas